requirements.txt
psycopg2-binary
python-dotenv
pyTelegramBotAPI
requests

